Move Contact Us config back into the client
