+++ app/shared/requiresSignin.ts
@@ -1,11 +1,7 @@
 import * as pathLib from "path";
 
 // To avoid security vulnerabilities do no add public paths that do not end in a slash
-const publicPaths = [
-  "/contact-us/",
-  "/api/contact-us-config/",
-  "/api/contact-us/"
-];
+const publicPaths = ["/contact-us/", "/api/contact-us/"];
 
 export const requiresSignin = (path: string) => {
   const normalizedPath = pathLib.normalize(path + "/");
